[Fix] add language feature-macro for <span> for detail::span
This commit fixes an issue with newer clang-versions support C++2a.
When compiling in C++17 mode, the compiler and libc++ do not expose
the <span> include. The logic for detail::span to include <span>
instead of providing an own implementation does only check if the
include <span> exists. It does, but it is not exposed.
Therefore the compilation fails without this fix.